% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_county.R
\name{stamp_county}
\alias{stamp_county}
\title{Title}
\usage{
stamp_county(
  mapping = NULL,
  data = reference_full,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  crs = "NAD27",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot() +
stamp_county()
}
